Изменил count на BigDecimal

+++ app/src/main/kotlin/com/itrocket/union/reserves/domain/entity/ReservesDomain.kt
@@ -3,6 +3,7 @@ package com.itrocket.union.reserves.domain.entity
 import android.os.Parcelable
 import com.itrocket.union.accountingObjects.domain.entity.ObjectInfoDomain
 import kotlinx.parcelize.Parcelize
+import java.math.BigDecimal
 
 @Parcelize
 data class ReservesDomain(
@@ -10,7 +11,7 @@ data class ReservesDomain(
     val title: String,
     val isBarcode: Boolean,
     val listInfo: List<ObjectInfoDomain>,
-    val itemsCount: Long,
+    val itemsCount: BigDecimal,
     val comment: String = "",
     val barcodeValue: String?,
     val nomenclatureId: String?,
