+++ union_sync_api/src/main/java/com/example/union_sync_api/entity/ReserveUpdateSyncEntity.kt
@@ -1,8 +1,10 @@
 package com.example.union_sync_api.entity
 
+import java.math.BigDecimal
+
 data class ReserveUpdateSyncEntity(
     val id: String,
-    val count: Long,
+    val count: BigDecimal,
     val locationId: String?,
     val userUpdated: String?,
     val dateUpdate: Long
